libstagefright:  Remove warnings
Clean up warnings and cause compilation error if new warnings are introduced.


Conflicts:

	media/libstagefright/AudioPlayer.cpp
	media/libstagefright/AwesomePlayer.cpp
	media/libstagefright/LPAPlayer.cpp
	media/libstagefright/LPAPlayerION.cpp
	media/libstagefright/LPAPlayerPMEM.cpp
	media/libstagefright/TunnelPlayer.cpp
	media/libstagefright/mpeg2ts/ESQueue.cpp

+++ media/libstagefright/ACodec.cpp
@@ -522,7 +522,7 @@ status_t ACodec::allocateOutputBuffersFromNativeWindow() {
     }
 
     int format = (def.format.video.eColorFormat == (OMX_COLOR_FORMATTYPE)QOMX_COLOR_FormatYUV420PackedSemiPlanar64x32Tile2m8ka)?
-                 HAL_PIXEL_FORMAT_YCbCr_420_SP_TILED : def.format.video.eColorFormat;
+                 (int) HAL_PIXEL_FORMAT_YCbCr_420_SP_TILED : (int) def.format.video.eColorFormat;
     if(def.format.video.eColorFormat == OMX_QCOM_COLOR_FormatYVU420SemiPlanar)
         format = HAL_PIXEL_FORMAT_YCrCb_420_SP;
     if(def.format.video.eColorFormat == QOMX_COLOR_FormatYVU420PackedSemiPlanar32m4ka)
@@ -1320,7 +1320,7 @@ void ACodec::sendFormatChange() {
             CHECK_LE(rect.nTop + rect.nHeight - 1, videoDef->nFrameHeight);
 
             int format = (def.format.video.eColorFormat == (OMX_COLOR_FORMATTYPE)QOMX_COLOR_FormatYUV420PackedSemiPlanar64x32Tile2m8ka)?
-                 HAL_PIXEL_FORMAT_YCbCr_420_SP_TILED : def.format.video.eColorFormat;
+                 (int) HAL_PIXEL_FORMAT_YCbCr_420_SP_TILED : (int) def.format.video.eColorFormat;
 
 
             if( mSmoothStreaming ) {
